Prevent segfault when responding to closed HTTP request.
LWS destroys the wsi when an HTTP request is aborted by the client. We
would, in an asynchronous case, still try to request writing the response
some time later using the destroyed wsi, which would bang.

We now use a std::shared_ptr to track the HttpRequest, and have a flag
within that is set when the request is aborted which will prevent the
responder from writing to the wsi.